+++ scripts/ci/index_ref_doc.py
@@ -44,8 +44,7 @@ for extension_name, exts in get_index_data()['extensions'].items():
 
     candidates_sorted = sorted(filtered_exts, key=lambda c: parse_version(c['metadata']['version']), reverse=True)
     chosen = candidates_sorted[0]
-    if extension_name != 'partnercenter':
-        ALL_TESTS.append((extension_name, chosen['downloadUrl'], chosen['filename']))
+    ALL_TESTS.append((extension_name, chosen['downloadUrl'], chosen['filename']))
 
 
 class TestIndexRefDocsMeta(type):
@@ -56,13 +55,15 @@ class TestIndexRefDocsMeta(type):
                 if dep_url.get(ext_name):
                     dep_file = get_whl_from_url(dep_url[ext_name][0], dep_url[ext_name][1], self.whl_dir)
                 else:
-                    dep_file = ''
+                    dep_file = None
                 ext_file = get_whl_from_url(ext_url, filename, self.whl_dir)
                 ref_doc_out_dir = os.path.join(REF_DOC_OUT_DIR, ext_name)
                 if not os.path.isdir(ref_doc_out_dir):
                     os.mkdir(ref_doc_out_dir)
                 script_args = [sys.executable, REF_GEN_SCRIPT, '--extension-file', ext_file, '--output-dir',
-                               ref_doc_out_dir, '--dependent-file', dep_file]
+                               ref_doc_out_dir]
+                if dep_file:
+                    script_args.extend(['--dependent-file', dep_file])
                 try:
                     check_call(script_args)
                 except CalledProcessError as e: