+++ src/amg/azext_amg/_help.py
@@ -36,11 +36,23 @@ helps['grafana show'] = """
 helps['grafana backup'] = """
     type: command
     short-summary: Backup an Azure Managed Grafana instance's content to an achive.
+    examples:
+        - name: backup dashboards under 2 folders (data sources are included to enable name remapping when restore dashboards to another worksapce)
+          text: |
+            az grafana backup -g MyResourceGroup -n MyGrafana -d c:\\temp --folders-to-include "Prod" "Compute Gateway" --components datasources dashboards folders
+        - name: backup dashboards and skip a few folders
+          text: |
+            az grafana backup -g MyResourceGroup -n MyGrafana -d c:\\temp  --folders-to-exclude General "Azure Monitor" --components datasources dashboards folders
+
 """
 
 helps['grafana restore'] = """
     type: command
     short-summary: Restore an Azure Managed Grafana instance from an achive.
+    examples:
+       - name: Restore dashboards. If they are under folders, include "folders" in the components list; use "--remap-data-sources" so CLI will update dashboards to point to same data sourceses at target workspace
+         text: |
+           az grafana restore -g MyResourceGroup -n MyGrafana --archive-file backup\\dashboards\\ServiceHealth-202307051036.tar.gz --components dashboards folders --remap-data-sources
 """
 
 helps['grafana update'] = """