+++ src/containerapp-preview/azext_containerapp_preview/__init__.py
@@ -6,7 +6,7 @@
 from azure.cli.core import AzCommandsLoader
 
 from azext_containerapp_preview._help import helps  # pylint: disable=unused-import
-from azext_containerapp_preview._utils import (_get_azext_containerapp_module)
+from azext_containerapp_preview._utils import (_get_azext_containerapp_module, auto_install_containerapp_extension_if_not_exist)
 
 
 class ContainerappPreviewCommandsLoader(AzCommandsLoader):
@@ -18,6 +18,7 @@ class ContainerappPreviewCommandsLoader(AzCommandsLoader):
             client_factory=None)
         super(ContainerappPreviewCommandsLoader, self).__init__(cli_ctx=cli_ctx,
                                                                 custom_command_type=containerapp_preview_custom)
+        auto_install_containerapp_extension_if_not_exist(self)
 
     def load_command_table(self, args):
         from azext_containerapp_preview.commands import load_command_table