+++ src/networkcloud/azext_networkcloud/aaz/latest/networkcloud/virtualmachine/_power_off.py
@@ -13,7 +13,7 @@ from azure.cli.core.aaz import *
 
 @register_command(
     "networkcloud virtualmachine power-off",
-    is_experimental=True,
+    is_preview=True,
 )
 class PowerOff(AAZCommand):
     """Power off the provided virtual machine.
@@ -23,9 +23,9 @@ class PowerOff(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2023-05-01-preview",
+        "version": "2023-07-01",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.networkcloud/virtualmachines/{}/poweroff", "2023-05-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.networkcloud/virtualmachines/{}/poweroff", "2023-07-01"],
         ]
     }
 
@@ -161,7 +161,7 @@ class PowerOff(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2023-05-01-preview",
+                    "api-version", "2023-07-01",
                     required=True,
                 ),
             }